Update Forecast.m
Update allows for multi-row or multi-column time series forecasting. It further allows for graphing of the results of such data, and it provides results as only the forecasted data, rounded to 3 decimal places (for use with corporate financial data, stated in thousands, converted to millions, whereby it would have 3 decimal places).